Travel-not
Wondering about getting a lift thing. It doesn’t really make sense to
say the impossibility of getting a lift is down to their condition.
It’s down to other people… Or is it to do with accessible transport -
e.g. for wheelchair users? Would ATW pay to have someone else’s car
adapted?